+++ shotmanager/features/cameraBG/__init__.py
@@ -19,9 +19,7 @@
 Camera BG
 """
 
-import bpy
-
-from shotmanager.config import config
+# from shotmanager.config import config
 
 # from .greasepencil_props import UAS_ShotManager_RenderGlobalContext, UAS_ShotManager_RenderSettings
 from . import cameraBG_operators